:sparkles: RegionInfoView 파일 분리
